some bugfixes and new components
started to add a very simple yet efficient reverb

fixed some issus with midi-file parsing
	wrong delta time for type-0 midis

refactored the midi parser
	added additional constants to the midi parser
	data handling class for better code readability

fixed simple-synth "moving" notes due to numerical instability

fixed wave-format issues
	added correct chunk parsing

added some consts ;)

refactoring of the LCD component of the master-target
	better re-use

"better" freq-spectrum display for the master target

fixed UI issue with knobs due to missing initializations

fixed control-right-click crash due to missing midi binder

started to add a simple phaser element

started to add a stereo expansion gate

+++ KSynth/test/TestMidi.cpp
@@ -20,7 +20,7 @@ TEST(Midi, parse) {
 	MidiFile mf;
 	MidiParser p(f, mf);
 
-	std::cout << "Version: " << mf.getVersion() << std::endl;
+	std::cout << "Version: " << (int)mf.getVersion() << std::endl;
 
 }
